refactor: indicators screen
