+++ src/modules/indicators/types/indicatorsTypes.ts
@@ -1,3 +1,8 @@
+export interface CriterionRange {
+  level: number
+  range: string
+}
+
 export interface Indicator {
   id: number
   name: string
@@ -6,6 +11,11 @@ export interface Indicator {
   unit: string
   minValue: number
   maxValue: number
+  example: string
+  purpose: string
+  ranges: CriterionRange[]
+  details: string
+  nullValue: string
 }
 
 export interface IndicatorResponse {
